"Further test cases added"

+++ CheckoutKata/ICheckout.cs
@@ -4,5 +4,5 @@ public interface ICheckout
 {
     public void Scan(string item);
     
-    public double GetTotalPrice();
+    public int GetTotalPrice();
 }
