+++ CheckoutKata/Product.cs
@@ -7,9 +7,9 @@ public class Product
     [Key]
     public string Sku { get; set; }
     
-    public double RegularPrice { get; set; }
+    public int RegularPrice { get; set; }
     
-    public double SpecialPrice { get; set; }
+    public int SpecialPrice { get; set; }
     
     public int QuantityRequiredForSpecialPrice { get; set; }
 }
